+++ sdk-java/src/main/java/ly/count/sdk/java/internal/RCDownloadCallback.java
@@ -13,4 +13,3 @@ public interface RCDownloadCallback {
      */
     void callback(RequestResult rResult, String error, boolean fullValueUpdate, Map<String, RCData> downloadedValues);
 }
-