+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 * Session update time duration increased to 60 seconds from 30 seconds.
 * Adding remaining request queue size information to every request.
 * Adding application version information to every request.
-* Added the remote config feature. Added consent for it "Config.Feature.RemoteConfig".
+* Added the remote config feature.
 * Remote Config module is accessible through "Countly::instance()::remoteConfig()" call.
 * A/B testing is added and also accessible through "Countly::instance()::remoteConfig()" call. 'testing' calls are not added.
 * Added configuration functions to configure Remote Config module on init:
@@ -14,6 +14,9 @@
   * 'enrollABOnRCDownload' to enroll A/B tests when remote config values downloaded
   * 'enableRemoteConfigAutomaticTriggers' to automatically download remote config values on init
   * 'remoteConfigRegisterGlobalCallback(RCDownloadCallback callback)' to register a remote config callback
+* Added the ability to set the user profile picture with an URL
+
+* Fixed a bug where it was not possible to send a profile picture with binary data
 
 ## 23.8.0